Adding a new folder called utils to save common code and also adding an auto calculate function to sum all the prices
