+++ src/components/Customer.jsx
@@ -1,7 +1,7 @@
 import data from "../data/data.json";
 import style from "./Customer.module.css";
+import { CustomerTableDetails } from "./CustomerTableDetails";
 import { Search } from "./Search";
-import { LuClipboardEdit } from "react-icons/lu";
 
 export const Customer = () => {
   return (
@@ -23,19 +23,9 @@ export const Customer = () => {
               </tr>
             </thead>
             <tbody>
-              {data.map((item) => {
+              {data.map((customer) => {
                 return (
-                  <tr key={item.id} className={style.tr}>
-                    <td className={style.td}>{item.Name}</td>
-                    <td className={style.td}>{item.Email}</td>
-                    <td className={style.td}>{item.address}</td>
-                    <td className={style.td}>{item.Phone}</td>
-                    <td className={style.td}>
-                      <button className="button is-info is-small">
-                        <LuClipboardEdit size={15} />
-                      </button>
-                    </td>
-                  </tr>
+                  <CustomerTableDetails key={customer.id} customer={customer} />
                 );
               })}
             </tbody>
