+++ src/components/Sidebar.module.css
@@ -40,6 +40,7 @@
     list-style: none;
     text-decoration: none;
     color: #2C3E50;
+
 }
 
 .footer {
@@ -54,9 +55,10 @@
 }
 
 
-.sidebar a:hover:not(.active) {
-    background-color: #5D6D7E;
-    transition: 1s ease;
+.linksection a:hover:not(.active) {
+    background-color: #D5D8DC;
+    transition: 0.5s ease;
+    text-transform: uppercase;
     color: white;
     border-radius: 5px;
 }
