+++ src/utils/DollarFormat.js
@@ -1,0 +1,4 @@
+export const DollarFormat = new Intl.NumberFormat("en-US", {
+  style: "currency",
+  currency: "USD",
+});
